restructuration of files (not all on main)
