+++ Drawable.py
@@ -1,0 +1,8 @@
+class Drawable:
+    
+    def __init__(self, surface, camera):
+        self.surface = surface
+        self.camera = camera
+        
+    def draw(self):
+        pass
