+++ Painter.py
@@ -1,0 +1,11 @@
+class Painter:
+
+    def __init__(self):
+        self.paintings = []
+
+    def add(self, drawable):
+        self.paintings.append(drawable)
+
+    def paint(self):
+        for drawing in self.paintings:
+            drawing.draw()
