+++ config.py
@@ -1,0 +1,8 @@
+import pygame
+
+SCREEN_WIDTH, SCREEN_HEIGHT = (1000,600)
+PLATFORM_SIZE = 4000
+SCREEN = pygame.display.set_mode((SCREEN_WIDTH,SCREEN_HEIGHT))
+GRID_SIZE = 50
+SIZE_OF_FOOD = 5
+NB_FOOD = PLATFORM_SIZE * 10
